Port `StaticMergeAllEventStream` to vertices
